+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Merged branches and closed issues
 
 - Schedules - Add `long_time_trace` and `long_time_trace_with_qubit` to perform  custom integrated time traces. (!878)
+- Qblox backend - Speedup execution by starting and stopping all armed sequencers in a Cluster via a single call. (!876)
 
 ## 0.19.0 (2024-04-03)